VerticalFragment completed
