+++ app/build.gradle
@@ -35,5 +35,6 @@ dependencies {
     compile 'com.firebase:firebase-client-android:2.5.2+'
     compile 'com.android.support:cardview-v7:+'
     compile 'com.github.moondroid.coverflow:library:1.0'
+    compile 'mobi.parchment:parchment:1.6.9@aar'
 
 }
