+++ app/src/main/java/com/kokushiseiya/tree/MainActivity.java
@@ -5,15 +5,7 @@ import android.app.FragmentTransaction;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.support.design.*;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v4.app.FragmentManager;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
-import android.view.Menu;
-import android.view.MenuItem;
 
 import com.firebase.client.Firebase;
 
@@ -41,6 +33,15 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onResume() {
         super.onResume();
+
+//        // Fragmentの動的追加
+//        android.app.FragmentManager fragmentManager = getFragmentManager();
+//        FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
+//        if (horizontalFragment == null) {
+//            horizontalFragment = new HorizontalFragment();
+//        }
+//        fragmentTransaction.replace(R.id.fragment_container, horizontalFragment);
+//        fragmentTransaction.commit();
 
         // Fragmentの動的追加
         android.app.FragmentManager fragmentManager = getFragmentManager();
