+++ app/src/main/res/drawable/edittext_style.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape android:shape="rectangle">
+            <!-- ボタンの線 -->
+            <stroke android:width="1dp" android:color="#838688" />
+            <!-- ボタンの色 -->
+            <solid android:color="#FFFFFF"></solid>
+            <!-- ボタンの角の丸さ -->
+            <corners android:radius="17dp" />
+        </shape>
+    </item>
+</selector>
+
