--- app/src/main/res/drawable/login_button_style.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <?xml version="1.0" encoding="utf-8"?>
-    <!-- ボタンが押されてないとき -->
-    <item android:state_pressed="false">
-        <shape android:shape="rectangle">
-            <!-- ボタンの線 -->
-            <stroke
-                android:width="1dp"
-                android:color="#838688" />
-            <!-- ボタンの色 -->
-            <solid android:color="#81C784"></solid>
-            <!-- ボタンの角の丸さ -->
-            <corners android:radius="17dp" />
-        </shape>
-    </item>
-
-    <!-- ボタンが押されたとき -->
-    <item android:state_pressed="true">
-        <!-- ボタンの線 -->
-        <stroke
-            android:width="1dp"
-            android:color="#838688" />
-        <shape android:shape="rectangle">
-            <!-- ボタンの色 -->
-            <solid android:color="#43A047"></solid>
-            <!-- ボタンの角の丸さ -->
-            <corners android:radius="17dp" />
-        </shape>
-    </item>
-</shape>
