--- app/src/main/res/layout/activity_login.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#c8e6c9">
-
-    <EditText
-        android:id="@+id/editText"
-        android:layout_width="261dp"
-        android:layout_height="53dp"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="129dp" />
-
-    <EditText
-        android:id="@+id/editText2"
-        android:layout_width="261dp"
-        android:layout_height="53dp"
-        android:layout_below="@id/editText"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="73dp" />
-
-    <Button
-        android:id="@+id/login_button"
-        android:layout_width="138dp"
-        android:layout_height="42dp"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="399dp"
-        android:background="@drawable/login_button_style"
-        android:text="New Button" />
-
-    <Button
-        android:id="@+id/singup_button"
-        android:layout_width="138dp"
-        android:layout_height="42dp"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="460dp"
-        android:background="@drawable/singup_button_style"
-        android:text="New Button" />
-
-    <ImageButton
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:src="@mipmap/ic_vpn_key_black"/>
-
-</RelativeLayout>
