+++ app/src/main/res/layout/fragment_horizontal.xml
@@ -1,24 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                xmlns:tools="http://schemas.android.com/tools"
                 xmlns:coverflow="http://schemas.android.com/apk/res-auto"
+                xmlns:tools="http://schemas.android.com/tools"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:paddingBottom="@dimen/activity_vertical_margin"
-                android:paddingLeft="@dimen/activity_horizontal_margin"
-                android:paddingRight="@dimen/activity_horizontal_margin"
-                android:paddingTop="@dimen/activity_vertical_margin"
                 tools:context=".MainActivity">
 
     <it.moondroid.coverflow.components.ui.containers.FeatureCoverFlow
-        android:id="@+id/ideaflow"
-        coverflow:coverHeight="150dp"
-        coverflow:coverWidth="100dp"
-        coverflow:maxScaleFactor="1.5"
+        android:id="@+id/coverflow"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        coverflow:coverHeight="196dp"
+        coverflow:coverWidth="202dp"
         coverflow:reflectionGap="0px"
         coverflow:rotationThreshold="0.5"
         coverflow:scalingThreshold="0.5"
-        coverflow:spacing="0.6"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        coverflow:shouldRepeat="true"
+        coverflow:spacing="0.6"/>
+
 </RelativeLayout>
