+++ app/src/main/res/layout/fragment_vertical.xml
@@ -1,20 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:coverflow="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
+                xmlns:tools="http://schemas.android.com/tools"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                tools:context=".MainActivity">
 
-    <it.moondroid.coverflow.components.ui.containers.FeatureCoverFlow
-        android:id="@+id/coverflow"
-        coverflow:coverHeight="196dp"
-        coverflow:coverWidth="202dp"
-        coverflow:reflectionGap="0px"
-        coverflow:rotationThreshold="0.5"
-        coverflow:scalingThreshold="0.5"
-        coverflow:spacing="0.6"
-        coverflow:shouldRepeat="true"
+
+    <ListView
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="match_parent"
+        android:id="@+id/listView"
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true"
+        android:divider="@null"
+        android:padding="10dp"
+        android:dividerHeight="20dp"
+        android:requiresFadingEdge="vertical"/>
+
 </RelativeLayout>
