+++ app/src/main/res/values/colors.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#3F51B5</color>
-    <color name="colorPrimaryDark">#303F9F</color>
-    <color name="colorAccent">#FF4081</color>
+    <color name="colorPrimary">#81C784</color>
+    <color name="colorPrimaryDark">#81C784</color>
+    <color name="colorAccent">#838688</color>
+    <color name="backgroundcolor">#C8E6C9</color>
 </resources>
